requirements.txt
blinker==1.7.0
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
export==0.2.1
firebase-admin==6.4.0
Flask==3.0.2
Flask-Login==0.6.3
Flask-Login-Multi==0.1.2
Flask-SQLAlchemy==3.1.1
gcloud==0.18.3
google-api-core==2.17.1
google-api-python-client==2.119.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.15.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.62.0
grpcio-status==1.62.0
gunicorn==21.2.0
httplib2==0.22.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jwcrypto==1.5.4
MarkupSafe==2.1.5
msgpack==1.0.7
oauth2client==4.1.3
packaging==23.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.20.0
PyJWT==2.8.0
pyparsing==3.1.1
Pyrebase4==4.7.1
python-jwt==4.1.0
requests==2.29.0
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.28
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==1.26.18
Werkzeug==3.0.1
